created notification class for managing notificaitons via slack and email
